feat: install & setup laravel sanctum

+++ app/Http/Controllers/Screencast/VideoController.php
@@ -57,6 +57,7 @@ class VideoController extends Controller
 
     public function destroy(Video $video)
     {
+        $this->authorize('delete', $video->playlist);
         $video->delete();
         return back();
     }
